Add support for `connection_scope` in params
Many identity providers recommend asking for permissions progressively
as per the requirements of your app, rather than all up-front.

Auth0 provides an option for requesting scopes for an identity provider
with the `connection_scope` parameter. Right now this can only be set
statically in the OmniAuth initializer.

This makes it hard to progressively request extra scopes.

A pattern for adding dynamic parameters to `authorize_params` already
exists and is used for `connection` and `prompt`.

We can extend this to add support for `connection_scope` too.

The values passed to `connection_scope` are often unique to the identity
provider and so will only be applied if they are supported by the
connection chosen by the user or the `connection` parameter.