Setup build matrix in CI (#116)
* allow GET requests in specs

This is not something anyone should do in production. But in tests, this
allows us to bypass CSRF.

* remove Gemfile.lock

So we're not constrained on a specific bundler version.

Also, Gemfile.lock won't be used when this gem is added as a dependency
anyway. If we require a specific version, it should be set in the
gemspec file.

* relax bundler version

1.9 is already pretty old. There's no reason not to be able to run a
  newer version

* setup ruby versions matrix

* upgrade listen

So it runs a version supported by ruby 3.0

* test ruby 3.0