Store plain Hash in session['authorize_params']
The `params` object that gets stored in `session['authorize_params']`
for token verification is an instance of `OmniAuth::Strategy::Options`,
an eventual subclass of `Hashie::Mash`. In reality, it's always treated
as a regular hash, and none of `Hashie::Mash`'s features are used.

Since values in the session get serialized into a session store (e.g.,
`Redis::Store` in our case), this `session['authorize_params]'` value is
getting serialized as an instance of `OmniAuth::Strategy::Options`. This
causes problems when code that doesn't have the
`OmniAuth::Strategy::Options` class loaded tries to load the session and
deserialize its contents. Since this `Options` class is unknown, the
code that's loading the session blows up.

This patch simply gets a regular hash from the `params` object before
storing it in the session. This way, we're just storing a plain Ruby
`Hash`, which is guaranteed to be available by any code that needs to
deserialize the session.